# openApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "cryptotracker.webapi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Integration": {
      "get": {
        "tags": [
          "Integration"
        ],
        "operationId": "GetIntegrations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Integration"
        ],
        "operationId": "AddIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIntegrationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIntegrationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddIntegrationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Integration/{id}/detail": {
      "get": {
        "tags": [
          "Integration"
        ],
        "operationId": "GetIntegrationDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Integration/{id}": {
      "put": {
        "tags": [
          "Integration"
        ],
        "operationId": "UpdateIntegration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Integration"
        ],
        "operationId": "DeleteIntegration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Integration/{id}/standing/days/{days}": {
      "get": {
        "tags": [
          "Integration"
        ],
        "operationId": "GetIntegrationStandingByDays",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "days",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "number",
                    "format": "double"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "number",
                    "format": "double"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "number",
                    "format": "double"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Integration/{id}/measuring": {
      "get": {
        "tags": [
          "Integration"
        ],
        "operationId": "GetMeasuringsByIntegration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyHoldingDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyHoldingDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyHoldingDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Integration"
        ],
        "operationId": "AddIntegrationMeasuring",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMeasuringDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMeasuringDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddMeasuringDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Integration"
        ],
        "operationId": "DeleteIntegrationMeasuring",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddIntegrationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddMeasuringDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AssetDto": {
        "required": [
          "assetType",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "assetType": {
            "$ref": "#/components/schemas/AssetType"
          },
          "isHidden": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AssetHoldingDto": {
        "required": [
          "asset",
          "integrationValues",
          "price",
          "totalAmount",
          "totalValue"
        ],
        "type": "object",
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetDto"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "totalValue": {
            "type": "number",
            "format": "double"
          },
          "integrationValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegrationAmount"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AssetType": {
        "enum": [
          "Fiat",
          "Crypto",
          "Stock",
          "ETF",
          "Commodity",
          "RealEstate"
        ],
        "type": "string"
      },
...
      "DailyHoldingDto": {
        "required": [
          "integrationId",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "integrationId": {
            "type": "string",
            "format": "uuid"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "$ref": "#/components/schemas/HoldingSource"
          }
        },
        "additionalProperties": false
      },
      "HoldingSource": {
        "enum": [
          "Sync",
          "Manual"
        ],
        "type": "string"
      },
      "IntegrationAmount": {
        "required": [
          "amount",
          "integration"
        ],
        "type": "object",
        "properties": {
          "integration": {
            "$ref": "#/components/schemas/IntegrationDto"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "IntegrationDetails": {
        "required": [
          "integration",
          "measurings"
        ],
        "type": "object",
        "properties": {
          "integration": {
            "$ref": "#/components/schemas/IntegrationDto"
          },
          "measurings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetHoldingDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegrationDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isHidden": {
            "type": "boolean"
          },
          "isManual": {
            "type": "boolean"
          },
          "lastSyncedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "currentValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateIntegrationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}